api/farm_health/requirements/local.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/local.txt requirements/local.in
#
alabaster==0.7.12         # via sphinx
appnope==0.1.0            # via ipython
argon2-cffi==18.3.0
atomicwrites==1.2.1       # via pytest
attrs==18.2.0             # via pytest
babel==2.6.0              # via sphinx
backcall==0.1.0           # via ipython
certifi==2018.10.15       # via requests
cffi==1.11.5              # via argon2-cffi
chardet==3.0.4            # via requests
coreapi==2.3.3
coreschema==0.0.4         # via coreapi
coverage==4.5.1
decorator==4.3.0          # via ipython, traitlets
defusedxml==0.5.0         # via python3-openid
django-allauth==0.38.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-extensions==2.1.3
django-model-utils==3.1.2
django-redis==4.9.0
django==2.1.2
djangorestframework==3.9.0
docutils==0.14            # via sphinx
factory-boy==2.11.1
faker==0.9.2              # via factory-boy
flake8==3.6.0
idna==2.7                 # via requests
imagesize==1.1.0          # via sphinx
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==7.1.1            # via ipdb
itypes==1.1.0             # via coreapi
jedi==0.13.1              # via ipython
jinja2==2.10              # via coreschema, sphinx
markupsafe==1.1.0         # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==4.3.0     # via pytest
mypy-extensions==0.4.1    # via mypy
mypy==0.641
oauthlib==2.1.0           # via requests-oauthlib
packaging==18.0           # via pytest-sugar, sphinx
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==5.3.0
pluggy==0.8.0             # via pytest
prompt-toolkit==2.0.7     # via ipython
psycopg2-binary==2.7.5
ptyprocess==0.6.0         # via pexpect
py==1.7.0                 # via pytest
pycodestyle==2.4.0        # via flake8
pycparser==2.19           # via cffi
pyflakes==2.0.0           # via flake8
pygments==2.2.0           # via ipython, sphinx
pyparsing==2.3.0          # via packaging
pytest-django==3.4.3
pytest-sugar==0.9.2
pytest==3.10.0
python-dateutil==2.7.5    # via faker
python-slugify==1.2.6
python3-openid==3.1.0     # via django-allauth
pytz==2018.7
redis==2.10.6
requests-oauthlib==1.0.0  # via django-allauth
requests==2.20.1          # via coreapi, django-allauth, requests-oauthlib, sphinx
six==1.11.0               # via argon2-cffi, django-coverage-plugin, django-extensions, faker, more-itertools, packaging, prompt-toolkit, pytest, python-dateutil, sphinx, traitlets
snowballstemmer==1.2.1    # via sphinx
sphinx==1.8.1
sphinxcontrib-websupport==1.1.0  # via sphinx
sqlparse==0.2.4           # via django-debug-toolbar
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.2       # via faker
traitlets==4.3.2          # via ipython
typed-ast==1.1.0          # via mypy
unidecode==1.0.22         # via python-slugify
uritemplate==3.0.0        # via coreapi
urllib3==1.24.1           # via requests
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.14.1
